Remove V19-05 datasets (#30)
* Mark V19-05 u, v & w fields datasets as inactive

This is in preparation for their removal.

* Mark V19-05 u, v & w 1mo fields datasets inactive

This is in preparation for their removal.

* Mark V19-05 biology fields datasets as inactive

This is in preparation for their removal.

* Mark V19-05 physics tracer fields datasets inactive

This is in preparation for their removal.

* Mark V19-05 aux vars & surface fields datasets inactive

This is in preparation for their removal.

* Mark V17-02 depth-avg seasonal fields datasets inactive

This is in preparation for their removal.

* Mark V17-02 bathymetry dataset as inactive

This is in preparation for their removal.

* Mark V17-02 mesh mask datasets as inactive

This is in preparation for their removal.

* Drop mentions of NEMO bathymetry in VHFR datasets

VHFR FVCOM datasets contain their own grid information. The
ubcSSnBathymetryV17-02 dataset was used in the calculation
of the boundary forcing fields for the FVCOM runs.

* Drop mention of NEMO bathymetry in wwatch3 datasets

wwatch3 datasets contain their own grid information. The
ubcSSnBathymetryV17-02 dataset was used in the calculation
of the surface current forcing field for the wwatch3 runs.

* Update rolling forecast datasets summary attrs

* Change georef dataset to ubcSSnBathymetryV21-08.
* Add explanation of lack of versioning of datasets due to rolling
  association with present near-real-time NEMO model
  configuration.

* Drop mention of V19-05 datasets from index page

* Improve line breaks in index page source

* Change diatoms to phytoplankton in ref example

* Drop Moore-Maley et al, 2016 from bio ref example

* Add zooplankton dataset ref example & Suchy etal 2023

* Fix typo in zooplankton ref example

* Delete V17-02 bathy & mesh mask datasets

* Delete V17-02 depth-avg seasonal fields dataset

* Delete V19-05 month-averaged fields datasets

* Delete V19-05 hour-averaged fields datasets

* Add bash scripts to flag active datasets to reload

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>