Fix bugs in GHA pytest workflow (#63)
* Use matrix.python-version in GHA pytest workflow

Replaced `inputs.python-version` with `matrix.python-version` in the
`check-datasets-xml`` workflow. This fixes a copy-paste error from the reusable
`pytest-with-coverage` workflow where the Python version is an input.

* Add date-based cache key to GHA pytest workflow

Introduced a missing step to generate a date-based cache key for downloads in
the `check-datasets-xml` workflow. This ensures that downloads are cached for 1
day. It is unclear how long the downloads were cached previously when the date
part of the cache key was empty.

* Add cache key for environment caching

Add a missing config item to set the environment cache key so
that the environment is cached for 1 day.